refactored translation that improves translation loading
